+++ src/main/java/roomescape/domain/reservationtime/ReservationTimeController.java
@@ -23,38 +23,6 @@ import roomescape.domain.reservationtime.dto.ReservationTimeResponse;
 public class ReservationTimeController {
 
     private final ReservationTimeService reservationTimeService;
-    private final AdminRequestValidator validator;
-
-    @GetMapping("/admin/times")
-    public ResponseEntity<List<ReservationTimeResponse>> getAllReservationTime(HttpServletRequest request) {
-        if (validator.isUnauthorized(request)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-        List<ReservationTimeResponse> response = reservationTimeService.getAllReservationTime();
-        return ResponseEntity.ok(response);
-    }
-
-    @PostMapping("/admin/times")
-    public ResponseEntity<TimeCreationResponse> createReservationTime(
-        HttpServletRequest httpServletRequest,
-        @RequestBody TimeCreationRequest createTimeRequest
-    ) {
-        if (validator.isUnauthorized(httpServletRequest)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-        createTimeRequest.validate();
-        TimeCreationResponse response = reservationTimeService.createReservationTime(createTimeRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
-    }
-
-    @DeleteMapping("/admin/times/{id}")
-    public ResponseEntity<Void> deleteReservationTime(HttpServletRequest request, @PathVariable Long id) {
-        if (validator.isUnauthorized(request)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-        reservationTimeService.deleteReservationTime(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
 
     @GetMapping("/times")
     public ResponseEntity<List<ReservationTimeAvailabilityResponse>> getReservationTimeAvailability(
