+++ src/main/java/roomescape/domain/theme/ThemeController.java
@@ -23,35 +23,6 @@ import roomescape.domain.theme.dto.ThemeResponse;
 public class ThemeController {
 
     private final ThemeService themeService;
-    private final AdminRequestValidator validator;
-
-    @GetMapping("/admin/themes")
-    public ResponseEntity<List<AdminThemeResponse>> getAllThemeForAdmin(HttpServletRequest request) {
-        if (validator.isUnauthorized(request)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-        List<AdminThemeResponse> response = themeService.getAllThemeForAdmin();
-        return ResponseEntity.ok(response);
-    }
-
-    @PostMapping("/admin/themes")
-    public ResponseEntity<ThemeCreationResponse> createTheme(@RequestBody ThemeCreationRequest createThemeRequest,
-        HttpServletRequest httpServletRequest) {
-        if (validator.isUnauthorized(httpServletRequest)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-        ThemeCreationResponse response = themeService.createTheme(createThemeRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
-    }
-
-    @DeleteMapping("/admin/themes/{id}")
-    public ResponseEntity<Void> deleteTheme(@PathVariable Long id, HttpServletRequest request) {
-        if (validator.isUnauthorized(request)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-        themeService.deleteTheme(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
 
     @GetMapping("/themes")
     public ResponseEntity<List<ThemeResponse>> getAllTheme() {
